encryption: split into RevisionLabelPodDeployer and StaticPodNodeProvider
